requirements.txt
transformers
accelerate
termcolor
bitsandbytes
inflection
sentence_transformers
python-arango
qdrant-client
openai
pandas
scipy
matplotlib
tqdm
requests
click
python-dotenv
mysql-connector-python
fpdf